backend-python/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
python-dotenv==1.0.1
psycopg[binary]==3.2.13
alembic==1.14.1
gunicorn==23.0.0
